move TypePromptResourceDetails to separate file to fix circular dependency
